added gaurd for rmDir

+++ config/wdio/wdio.conf.js
@@ -128,8 +128,8 @@ const config = {
   before() {
     if (runLightHouse) {
       generateSessionToken();
-      fs.rmdirSync('report//performance-report.json');
-      fs.rmdirSync('report//performance-report.html');
+      if (!fs.existsSync('report//performance-report.json')) fs.rmdirSync('report//performance-report.json');
+      if (!fs.existsSync('report//performance-report.html')) fs.rmdirSync('report//performance-report.html');
     }
   },
